requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
supabase==2.18.1
google-generativeai==0.8.5
python-dotenv==1.1.1
httpx[http2]==0.28.1
pydantic==2.11.7
python-multipart==0.0.9
openai==1.100.2
# Dependencies for other services that might be co-located or for utility scripts
pydub==0.25.1
# Testing dependencies
pytest-asyncio==0.23.8
pytest-mock==3.14.0
prometheus-fastapi-instrumentator==7.0.0
